+++ examples/nextjs/pages/index.tsx
@@ -3,8 +3,12 @@ import Head from 'next/head';
 import styles from '../styles/Home.module.scss';
 
 const EXAMPLE_ROUTES = {
-  minimal: { title: 'Minimal example', href: () => `/minimal` },
-  simple: { title: 'Simple example', href: () => `/simple` },
+  voiceAssistant: {
+    title: 'AI Voice Assistant example',
+    href: () => `/voice-assistant`,
+  },
+  minimal: { title: 'VideoConference example with minimal code', href: () => `/minimal` },
+  simple: { title: 'Simple custom setup example', href: () => `/simple` },
   audioOnly: {
     title: 'Audio only example',
     href: () => `/audio-only`,
@@ -18,7 +22,7 @@ const EXAMPLE_ROUTES = {
     href: () => `/clubhouse`,
   },
   processors: {
-    title: 'Minimal example with background blur',
+    title: 'Example usage of @livekit/track-processors for background blur',
     href: () => `/processors`,
   },
 } as const;