+++ packages/react/src/prefabs/index.ts
@@ -4,3 +4,4 @@ export { VideoConference, type VideoConferenceProps } from './VideoConference';
 export { ControlBar, type ControlBarProps, type ControlBarControls } from './ControlBar';
 export { MediaDeviceMenu, type MediaDeviceMenuProps } from './MediaDeviceMenu';
 export { AudioConference, type AudioConferenceProps } from './AudioConference';
+export * from './VoiceAssistantControlBar';